redo cooldowns

+++ starboard/core/starboards.py
@@ -266,7 +266,7 @@ async def _add_reactions(
 
 
 EDIT_COOLDOWN: FixedCooldown[int] = FixedCooldown(
-    CONFIG.edit_cooldown_period, CONFIG.edit_cooldown_cap
+    *CONFIG.guild_message_edit_cooldown
 )
 
 
@@ -278,7 +278,8 @@ async def _edit(
     embeds: list[hikari.Embed] | None,
     author_id: int,
 ) -> None:
-    if EDIT_COOLDOWN.update_ratelimit(message.id):
+    assert message.guild_id
+    if EDIT_COOLDOWN.update_ratelimit(message.guild_id):
         return
 
     if message.author.id != bot.me.id:
@@ -301,9 +302,18 @@ async def _edit(
         )
 
 
+DELETE_COOLDOWN: FixedCooldown[int] = FixedCooldown(
+    *CONFIG.guild_message_delete_cooldown
+)
+
+
 async def _delete(
     bot: Bot, config: StarboardConfig, message: hikari.Message
 ) -> None:
+    assert message.guild_id
+    if DELETE_COOLDOWN.update_ratelimit(message.guild_id):
+        return
+
     if message.author.id == bot.me.id:
         return await message.delete()
 
@@ -318,6 +328,11 @@ async def _delete(
                 await message.delete()
 
 
+SEND_COOLDOWN: FixedCooldown[int] = FixedCooldown(
+    *CONFIG.guild_message_send_cooldown
+)
+
+
 async def _send(
     bot: Bot,
     config: StarboardConfig,
@@ -325,6 +340,9 @@ async def _send(
     embeds: list[hikari.Embed] | None,
     author_id: int,
 ) -> hikari.Message | None:
+    if SEND_COOLDOWN.update_ratelimit(config.starboard.guild_id):
+        return None
+
     webhook = await _webhook(bot, config)
 
     if webhook and config.use_webhook:
